♻️  Split Emoji into own module
