+++ src/emojis.rs
@@ -1,6 +1,6 @@
 use std::{error::Error, fs::read_to_string, io, path::PathBuf};
 
-use regex::Regex;
+use crate::emoji::Emoji;
 
 #[derive(serde::Deserialize, Debug)]
 pub struct Emojis {
@@ -40,34 +40,3 @@ pub fn cache_dir() -> Result<PathBuf, Box<dyn Error>> {
 const CACHE_DIR: &str = "gimoji";
 const EMOJI_CACHE_FILE: &str = "emojis.json";
 const EMOJI_URL: &str = "https://gitmoji.dev/api/gitmojis";
-
-#[derive(serde::Deserialize, Debug)]
-pub struct Emoji {
-    code: String,
-    description: String,
-    emoji: String,
-    entity: String,
-    name: String,
-}
-
-impl Emoji {
-    pub fn contains(&self, pattern: &Regex) -> bool {
-        pattern.is_match(&self.code)
-            || pattern.is_match(&self.description)
-            || pattern.is_match(&self.emoji)
-            || pattern.is_match(&self.entity)
-            || pattern.is_match(&self.name)
-    }
-
-    pub fn code(&self) -> &str {
-        self.code.as_ref()
-    }
-
-    pub fn description(&self) -> &str {
-        self.description.as_ref()
-    }
-
-    pub fn emoji(&self) -> &str {
-        self.emoji.as_ref()
-    }
-}
