+++ src/selection_view.rs
@@ -6,7 +6,7 @@ use ratatui::{
 };
 use regex::RegexBuilder;
 
-use crate::emojis::Emoji;
+use crate::emoji::Emoji;
 
 pub struct SelectionView {
     emojis: Vec<Emoji>,
